fix(rebase): bring TASK-141 forward onto beta after PR #130's collateral close
PR #130 was auto-closed on 2026-08-29 when `beta` was briefly deleted and
recreated. It was never reviewed and never rejected, and the six fixes in it
have sat unmerged since.

Every finding was re-verified against current `beta` before rebasing rather
than assumed still present. Five of six are still live. The sixth -- JOIN's
elevation handling -- was fixed upstream by issue #373, so this branch now
yields to it entirely: six of the eleven conflicts were that overlap, and all
six resolve in upstream's favour. Upstream's JOIN also carries REQ-316/ADR-047
bulges and REQ-325/ADR-053 tilted-arc planes that this branch predates.

Three adaptations the eleven days made necessary:

  ADR-054 widened the coordinate stores to double underneath this work, so
  three records reading from them into float fields became ill-formed
  narrowing conversions. Each is now an explicit cast with the reason written
  at the declaration -- both sit downstream of interfaces that are still
  float-in (`readVert`, `CadCoord::WorldFromLocal`), so widening the record
  alone would move the narrowing one line down and buy nothing. Widening those
  paths is ADR-054's own audit to make.

  Issue #264 Stage 1 retired the standalone `.gs` format; regression-issue08
  now opens the `.dwg` every other surface transcript uses.

  QUICKSELECT gained BlockRef support upstream after this branch moved
  ExecuteQuickSelect to the command layer; the relocated copy carries those
  three additions forward rather than silently reverting them.

Issue 03 re-proven to bite rather than asserted: disabling the pre-scan turns
its transcript red on a value difference, losing up to 0.108 ft on a
state-plane easting. That is 54x REQ-101's +/-0.002 ft -- when this task was
written the tolerance was +/-0.01 and the same error was 10.8x, so ADR-054 did
not make the finding stale, it made it five times worse.

ctest 1367/1367 green, up from the 743/743 the task was originally verified
against.

+++ tests/headless/HeadlessDriver.cpp
@@ -1385,6 +1385,12 @@ bool ExecuteStep(Run& run, const std::string& raw, int sourceLine) {
       kind = UpperAscii(kind);
 
       // Collect the entity's vertices as (localX, localY, absoluteZ), whatever store it lives in.
+      //
+      // `float`, matching `CadCoord::WorldFromLocal`'s own parameters below: the stores are
+      // `double` since ADR-054 Phase A, but the local->world conversion this oracle has to go
+      // through is still float-in, so widening the collection alone would buy no precision and
+      // would only move the narrowing one line further down. The casts are explicit so that is a
+      // stated limit of the oracle rather than an accident of the store's type.
       std::vector<std::array<float, 3>> verts;
       std::string why;
       if (kind == "LINE") {
@@ -1393,8 +1399,8 @@ bool ExecuteStep(Run& run, const std::string& raw, int sourceLine) {
           why = "no line at index " + std::to_string(ei) + " (there are " +
                 std::to_string(run.st.userLinesFlat.size() / 6) + ")";
         else
-          verts = {{{run.st.userLinesFlat[base], run.st.userLinesFlat[base + 1], run.st.userLinesFlat[base + 2]}},
-                   {{run.st.userLinesFlat[base + 3], run.st.userLinesFlat[base + 4], run.st.userLinesFlat[base + 5]}}};
+          verts = {{{static_cast<float>(run.st.userLinesFlat[base]), static_cast<float>(run.st.userLinesFlat[base + 1]), static_cast<float>(run.st.userLinesFlat[base + 2])}},
+                   {{static_cast<float>(run.st.userLinesFlat[base + 3]), static_cast<float>(run.st.userLinesFlat[base + 4]), static_cast<float>(run.st.userLinesFlat[base + 5])}}};
       } else if (kind == "POLYLINE") {
         const size_t n = PolylineCountOf(run.st);
         if (ei < 0 || static_cast<size_t>(ei) >= n)
@@ -1405,8 +1411,9 @@ bool ExecuteStep(Run& run, const std::string& raw, int sourceLine) {
           for (int v = b; v < e; ++v) {
             const size_t o = static_cast<size_t>(v) * 3;
             if (o + 2 < run.st.userPolylineVerts.size())
-              verts.push_back({{run.st.userPolylineVerts[o], run.st.userPolylineVerts[o + 1],
-                                run.st.userPolylineVerts[o + 2]}});
+              verts.push_back({{static_cast<float>(run.st.userPolylineVerts[o]),
+                                static_cast<float>(run.st.userPolylineVerts[o + 1]),
+                                static_cast<float>(run.st.userPolylineVerts[o + 2])}});
           }
         }
       } else if (kind == "CIRCLE") {
@@ -1415,15 +1422,15 @@ bool ExecuteStep(Run& run, const std::string& raw, int sourceLine) {
           why = "no circle at index " + std::to_string(ei) + " (there are " +
                 std::to_string(run.st.userCirclesCxCyZR.size() / 4) + ")";
         else  // centre + elevation; the radius is not a coordinate
-          verts = {{{run.st.userCirclesCxCyZR[base], run.st.userCirclesCxCyZR[base + 1],
-                     run.st.userCirclesCxCyZR[base + 2]}}};
+          verts = {{{static_cast<float>(run.st.userCirclesCxCyZR[base]), static_cast<float>(run.st.userCirclesCxCyZR[base + 1]),
+                     static_cast<float>(run.st.userCirclesCxCyZR[base + 2])}}};
       } else if (kind == "ARC") {
         if (ei < 0 || static_cast<size_t>(ei) >= run.st.userArcs.size())
           why = "no arc at index " + std::to_string(ei) + " (there are " +
                 std::to_string(run.st.userArcs.size()) + ")";
         else {
           const CadArc& a = run.st.userArcs[static_cast<size_t>(ei)];
-          verts = {{{a.cx, a.cy, a.z}}};
+          verts = {{{static_cast<float>(a.cx), static_cast<float>(a.cy), static_cast<float>(a.z)}}};
         }
       } else {
         Fail(run, "parse", "EXPECT " + what + ": unknown kind " + kind + " (LINE/POLYLINE/CIRCLE/ARC)",
